requirements.txt
numpy~=1.23.3
pandas~=1.5.1
hydra-core~=1.1.0
omegaconf~=2.1.1
scipy~=1.9.1
scikit-learn~=1.1.2
tqdm~=4.64.1
pyg~=2.0.4
pytorch~=1.11.0
